moves shared exceptions and tests into separate lib/folder
