+++ lib/src/exceptions.dart
@@ -41,18 +41,3 @@ class EntityDescriptionMissing extends Error {
   String toString() =>
       "EntityDescription missing: Entity ${_clazz} is not descriped in ENTITY_MAP.";
 }
-
-class NullObjectError extends Error {
-
-  String toString() => "Object must not be null";
-
-}
-
-class UnknownIdentifierError extends Error {
-  final String _identifier;
-
-  UnknownIdentifierError(this._identifier);
-
-  String toString() =>
-      "Type for identifier '${_identifier}' unknown. Use addIdentifier to register type information.";
-}
